feat: add highlight for keywords
