+++ .github/workflows/ci.yml
@@ -101,6 +101,12 @@ jobs:
           files: lcov.info
           fail_ci_if_error: true
 
+      - name: Install cargo hack
+        uses: taiki-e/install-action@cargo-hack
+
+      - name: run cargo check with all possible feature set
+        run: cargo hack check --feature-powerset
+
   # Run tests on Linux, macOS, and Windows
   # On both Rust stable and Rust nightly
   test:
